refactor(chess/domain/board): 체스판 초기화 담당 클래스 BoardInitializer 추가
- 초기 기물 위치를 알고, 그에 맞게 기물을 세팅하여 Board 생성
- Board 사용하는 Test들의 fixture 추가
