+++ src/main/java/chess/domain/gamestate/Ready.java
@@ -2,6 +2,7 @@ package chess.domain.gamestate;
 
 import chess.domain.Camp;
 import chess.domain.board.Board;
+import chess.domain.board.BoardInitializer;
 import chess.domain.board.Position;
 
 public class Ready implements State {
@@ -12,7 +13,7 @@ public class Ready implements State {
 
     @Override
     public State start() {
-        return new Running(new Board());
+        return new Running(BoardInitializer.get());
     }
 
     @Override
